backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
textblob
openai
redis
rq
sentence-transformers
faiss-cpu
python-dotenv
